Extract function from runAndroid and add tests for runOnAllDevices.
