+++ .gitignore
@@ -1,5 +1,6 @@
 .DS_Store
 node_modules
 .vscode
+.idea
 package-lock.json
 npm-debug.log
